debugging

+++ models/finetune.py
@@ -41,7 +41,7 @@ def get_ft(train_dataset, config, device, prompt_template="a photo of [attr] [ob
         None,
         prompt_template=prompt_template,
         device=device,
-        dtype=torch.float16,
+        dtype=torch.float32,
         enable_pos_emb=True,
     )
 
@@ -50,7 +50,7 @@ def get_ft(train_dataset, config, device, prompt_template="a photo of [attr] [ob
     )
     if config.amp:
         ft, optimizer = amp.initialize(
-            ft.float(), optimizer, opt_level="O2",
+            ft, optimizer, opt_level="O2",
             keep_batchnorm_fp32=True, loss_scale="dynamic"
         )
